Fargate: move API input validation logic to validation.go
... in order to keep such logic all in one unique file.
