+++ pkg/apis/eksctl.io/v1alpha5/types.go
@@ -14,7 +14,6 @@ import (
 	"github.com/aws/aws-sdk-go/service/iam/iamiface"
 	"github.com/aws/aws-sdk-go/service/ssm/ssmiface"
 	"github.com/aws/aws-sdk-go/service/sts/stsiface"
-	"github.com/pkg/errors"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/runtime"
 )
@@ -796,31 +795,6 @@ type FargateProfile struct {
 	Subnets []string `json:"subnets,omitempty"`
 }
 
-// ReservedProfileNamePrefix defines the Fargate profile name prefix reserved
-// for AWS, and which therefore, cannot be used by users. AWS' API should
-// reject the creation of profiles starting with this prefix, but we eagerly
-// validate this client-side.
-const ReservedProfileNamePrefix = "eks-"
-
-// Validate validates this FargateProfile object.
-func (fp FargateProfile) Validate() error {
-	if fp.Name == "" {
-		return errors.New("invalid Fargate profile: empty name")
-	}
-	if strings.HasPrefix(fp.Name, ReservedProfileNamePrefix) {
-		return fmt.Errorf("invalid Fargate profile \"%s\": name should NOT start with \"%s\"", fp.Name, ReservedProfileNamePrefix)
-	}
-	if len(fp.Selectors) == 0 {
-		return fmt.Errorf("invalid Fargate profile \"%s\": no profile selector", fp.Name)
-	}
-	for i, selector := range fp.Selectors {
-		if err := selector.Validate(); err != nil {
-			return errors.Wrapf(err, "invalid Fargate profile \"%s\": invalid profile selector at index #%v", fp.Name, i)
-		}
-	}
-	return nil
-}
-
 // FargateProfileSelector defines rules to select workload to schedule onto Fargate.
 type FargateProfileSelector struct {
 	// Namespace is the Kubernetes namespace from which to select workload.
@@ -828,12 +802,4 @@ type FargateProfileSelector struct {
 	// +optional
 	// Labels are the Kubernetes label selectors to use to select workload.
 	Labels map[string]string `json:"labels,omitempty"`
-}
-
-// Validate validates this FargateProfileSelector object.
-func (fps FargateProfileSelector) Validate() error {
-	if fps.Namespace == "" {
-		return errors.New("empty namespace")
-	}
-	return nil
 }
